==> 1_feature_backfill.ipynb <==
!pip install -U hopsworks --quiet

import numpy as np
import pandas as pd

#import transactions_fraud

#import great_expectations
# Mute warnings
import warnings
warnings.filterwarnings("ignore")
import pandas as pd
import kagglehub
import shutil

from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta
from sentence_transformers import SentenceTransformer

Download csv file from kaggle, then move it

path = kagglehub.dataset_download("gpreda/bbc-news")
print("Path to dataset files:", path) #"/home/jupyter/.cache/kagglehub/datasets/gpreda/bbc-news/versions/100" 
shutil.move(path, './')

Extract the data

def get_backfill():
    file_path = './bbc_news.csv'
    df = pd.read_csv(file_path)
    df = df[['pubDate', 'title']]
    df['pubDate'] = pd.to_datetime(df['pubDate'], format='%a, %d %b %Y %H:%M:%S GMT')
    # as 'YYYY-MM-DD'
    df['pubDate'] =  pd.to_datetime(df['pubDate'].dt.strftime('%Y-%m-%d'))
    df2 = df[df['pubDate'] > pd.to_datetime('2023-12-29')]
    print(df2.head())
    print(df.tail())
    print(len(df))
    return df2
df_headlines = get_backfill()

Get the embedding model

emb_model = SentenceTransformer('all-MiniLM-L6-v2')

KEY = "wOWQmzzHeITT5wlJ.FkOButuQ3XpDXSUO1LnTuaNfD7SPWolfF1hateoistsLFFsBF7upULn5z6qKPOIB"
project = hopsworks.login(api_key_value=KEY)

embeddings = emb_model.encode(df_headlines['title'].tolist())

print(embeddings.shape)

Write the embeddings, headlines and dates to hopsworks

df = pd.DataFrame(embeddings, columns=[f"dim_{i}" for i in range(384)])
print(df)
df["headlines"] = df_headlines['title'].tolist()
df["date"] = df_headlines['pubDate'].tolist()
feature_group = fs.get_or_create_feature_group(
    name="headlinesemb",
    version=1,
    description="Feature group for 384-dimensional vectors",
    primary_key=["headlines"], 
    online_enabled=True 
)
feature_group.insert(df, write_options={"wait_for_job": True})
print("384-dimensional vectors successfully written to Hopsworks Feature Store.")
