==> 2_feature_pipeline.ipynb <==
!pip install -U hopsworks --quiet

import requests
import pandas as pd
from datetime import datetime
from sentence_transformers import SentenceTransformer
import hopsworks

def get_headlines():
    response = requests.get('https://bbc-api.vercel.app/news?lang=english')
    news_data = response.json()
    title_set = set()
    for article in news_data['Latest']:
        title_set.add(article['title'])
    return list(title_set)

def df_time_headlines(headlines_today):
    df = pd.DataFrame(headlines_today, columns=["title"])
    df["pubdate"] = pd.to_datetime(datetime.today().strftime('%Y-%m-%d'))
    df = df[["pubdate", "title"]]
    return df

headlines_today = get_headlines()
today_df = df_time_headlines(headlines_today)
print(today_df)

emb_model = SentenceTransformer('all-MiniLM-L6-v2')
embeddings = emb_model.encode(today_df['title'].tolist())
df = pd.DataFrame(embeddings, columns=[f"dim_{i}" for i in range(384)])
df["headlines"] = today_df['title'].tolist()
df["date"] = today_df['pubdate'].tolist()
print(df)
assert len(df) >=1, 'No new headlines collected'

KEY = "wOWQmzzHeITT5wlJ.FkOButuQ3XpDXSUO1LnTuaNfD7SPWolfF1hateoistsLFFsBF7upULn5z6qKPOIB"
project = hopsworks.login(api_key_value=KEY)
fs = project.get_feature_store()
feature_group = fs.get_or_create_feature_group(
    name="headlinesemb",
    version=1,
    description="Feature group for 384-dimensional vectors",
    primary_key=["headlines"], 
    online_enabled=True 
)
feature_group.insert(df, write_options={"wait_for_job": True})
