==> 3_training_pipeline.ipynb <==
from sklearn.decomposition import PCA
import hopsworks
import pandas as pd
from PIL import Image
import matplotlib.pyplot as plt
import numpy as np
from sentence_transformers import SentenceTransformer
from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score
from sklearn.feature_extraction.text import CountVectorizer
from collections import Counter

from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

project = hopsworks.login(api_key_value="wOWQmzzHeITT5wlJ.FkOButuQ3XpDXSUO1LnTuaNfD7SPWolfF1hateoistsLFFsBF7upULn5z6qKPOIB")
fs = project.get_feature_store() 
headline_fg = fs.get_feature_group(
    name='headlinesemb', 
    version=1,
)
headline_df = headline_fg.read()
headline_df = headline_df.sort_values(by='date')

print(headline_df.head())
print(headline_df.tail())

def df_between_dates(df, start_date, end_date):
    return df[(df['date'] >= start_date) & (df['date'] <= end_date)]

now = datetime.now()
past_year = now + relativedelta(years=-1)
#df_week = df_between_dates(headline_df, '2024-11-01', '2024-12-3')
df = df_between_dates(headline_df, past_year.strftime("%Y-%m-%d"), now.strftime("%Y-%m-%d"))

emb_model = SentenceTransformer('all-MiniLM-L6-v2')

headlines= df['headlines'].tolist()
embeddings = df[[f"dim_{i}" for i in range(384)]].to_numpy()#emb_model.encode(headlines[key])

def training(embeddings, headlines, n_clusters=3):
    kmeans = KMeans(n_clusters=n_clusters, random_state=22)
    clusters = kmeans.fit(embeddings) # 1, 2 or 3... for each headline
    pca = PCA(n_components=2)
    pca.fit(embeddings)
    return kmeans, pca
n_clusters=5
kmeans,pca = training(embeddings, headlines, n_clusters)

emb =  emb_model.encode(['police', 'Trump'])
clusters = kmeans.predict(emb)
print(clusters)

import pickle
with open("pca_model.pkl", "wb") as pca_file:
    pickle.dump(pca, pca_file)

with open("kmeans_model.pkl", "wb") as kmeans_file:
    pickle.dump(kmeans, kmeans_file)

mr = project.get_model_registry()

pca_model = mr.python.create_model(
    name="PCA_Model",
    description="PCA model for dimensionality reduction"
)
kmeans_model = mr.python.create_model(
    name="KMeans_Model",
    description="KMeans clustering model"
)
pca_model.save('pca_model.pkl')
kmeans_model.save('kmeans_model.pkl')

Look at size

print(embeddings.shape)
print(len(headlines))

#size mb, headlines 1000 times larger
import sys
print(sys.getsizeof(embeddings)/ (1024**2))
print(sys.getsizeof(headlines)/ (1024**2))
